Moved models in model subdirectory/namespace
