+++ src/dart/simple_arm.cpp
@@ -61,8 +61,8 @@
 #include <limbo/opt/cmaes.hpp>
 
 #include <blackdrops/blackdrops.hpp>
-#include <blackdrops/gp_model.hpp>
 #include <blackdrops/model/gp/kernel_lf_opt.hpp>
+#include <blackdrops/model/gp_model.hpp>
 #include <blackdrops/system/dart_system.hpp>
 
 #include <blackdrops/policy/nn_policy.hpp>
@@ -346,7 +346,7 @@ int main(int argc, char** argv)
 
     using GP_t = limbo::model::MultiGP<Params, limbo::model::GP, kernel_t, mean_t, limbo::model::multi_gp::ParallelLFOpt<Params, blackdrops::model::gp::KernelLFOpt<Params>>>;
 
-    using MGP_t = blackdrops::GPModel<Params, GP_t>;
+    using MGP_t = blackdrops::model::GPModel<Params, GP_t>;
 
     blackdrops::BlackDROPS<Params, MGP_t, SimpleArm, blackdrops::policy::NNPolicy<PolicyParams>, policy_opt_t, RewardFunction> arm_system;
